Finalising product and category CRUD

+++ src/main/java/com/redaeilco/ecommerce/service/CategoryService.java
@@ -1,7 +1,9 @@
 package com.redaeilco.ecommerce.service;
 
 import com.redaeilco.ecommerce.model.Category;
+import com.redaeilco.ecommerce.model.User;
 import com.redaeilco.ecommerce.repository.CategoryRepository;
+import com.redaeilco.ecommerce.repository.UserRepository;
 
 import jakarta.persistence.EntityNotFoundException;
 
@@ -16,7 +18,14 @@ public class CategoryService {
     @Autowired
     private CategoryRepository categoryRepository;
 
+    @Autowired
+    private UserRepository userRepository;
+
+    @Autowired
+    private JWTService jwtService;
+
     public List<Category> getAllCategories() {
+        System.out.println("Fetching all categories");
         return categoryRepository.findAll();
     }
 
@@ -26,18 +35,31 @@ public class CategoryService {
             .orElseThrow(() -> new EntityNotFoundException("Category not found with ID: " + id));
     }
 
-    public Category createCategory(Category category) {
+    public Category createCategory(Category category, String token) {
+        int userId = jwtService.extractUserId(token);
+        User user = userRepository.findById(userId).get();
+        category.setCreatedBy(user);
         return categoryRepository.save(category);
     }
 
-    public Category updateCategory(int id, Category categoryDetails) {
+    public Category updateCategory(int id, Category categoryDetails, String token) {
         Category category = getCategoryById(id);
+        int userId = jwtService.extractUserId(token);
+        User user = userRepository.findById(userId).get();
+        if (user != category.getCreatedBy()) {
+            throw new RuntimeException("You can only update your own categories.");
+        }
         category.setName(categoryDetails.getName());
         return categoryRepository.save(category);
     }
 
-    public void deleteCategory(int id) {
-        getCategoryById(id); // will throw an exception if the category is not found
+    public void deleteCategory(int id, String token) {
+        Category category = getCategoryById(id);
+        int userId = jwtService.extractUserId(token);
+        User user = userRepository.findById(userId).get();
+        if (user != category.getCreatedBy()) {
+            throw new RuntimeException("You can only delete your own categories.");
+        }
         categoryRepository.deleteById(id);
     }
 }
